+++ lib/templates/di_template.dart
@@ -1,8 +1,12 @@
 import 'package:packed/utils/utils.dart';
 
+/// A template class for Dependency Injection files.
 class DiTemplate {
-  static String diTemplate(String name) {
+  /// Returns the template for a DI registration file for the given [name].
+  static String diTemplate(String name, {bool isBloc = false}) {
     final pascal = Utils.pascal(name);
+    final type = isBloc ? 'Bloc' : 'Cubit';
+    final folder = isBloc ? 'bloc' : 'cubit';
 
     return '''
 import 'package:get_it/get_it.dart';
@@ -10,13 +14,13 @@ import '../domain/repositories/${name}_repository.dart';
 import '../domain/usecases/get_${name}_usecase.dart';
 import '../data/datasources/${name}_remote_datasource.dart';
 import '../data/repositories/${name}_repository_impl.dart';
-import '../presentation/cubit/${name}_cubit.dart';
+import '../presentation/$folder/${name}_$folder.dart';
 
 final sl = GetIt.instance;
 
 Future<void> init${pascal}Feature() async {
-  // Cubit
-  sl.registerFactory(() => ${pascal}Cubit(get${pascal}UseCase: sl()));
+  // $type
+  sl.registerFactory(() => ${pascal}$type(get${pascal}UseCase: sl()));
 
   // Usecases
   sl.registerLazySingleton(() => Get${pascal}UseCase(sl()));