+++ lib/templates/repository_template.dart
@@ -6,11 +6,11 @@ class RepositoryTemplate {
 
     return '''
 import 'package:dartz/dartz.dart';
-import '../../core/error/failures.dart';
+import '../../../../core/error/failures.dart';
 import '../entities/${name}_entity.dart';
 
 abstract class ${pascal}Repository {
-  Future<Either<Failure, ${pascal}Entity>> get${pascal}();
+  Future<Either<Failure, ${pascal}Entity>> get$pascal();
 }
 ''';
   }