+++ lib/templates/usecase_template.dart
@@ -7,8 +7,8 @@ class UsecaseTemplate {
 
     return '''
 import 'package:dartz/dartz.dart';
-import '../../core/error/failures.dart';
-import '../../core/usecases/usecase.dart';
+import '../../../../core/error/failures.dart';
+import '../../../../core/usecases/usecase.dart';
 import '../entities/${featureName}_entity.dart';
 import '../repositories/${featureName}_repository.dart';
 
@@ -19,7 +19,7 @@ class ${pascal}UseCase implements UseCase<${featurePascal}Entity, NoParams> {
 
   @override
   Future<Either<Failure, ${featurePascal}Entity>> call(NoParams params) async {
-    return await repository.get${featurePascal}();
+    return await repository.get$featurePascal();
   }
 }
 ''';