+++ lib/templates/view_template.dart
@@ -1,15 +1,17 @@
 import 'package:packed/utils/utils.dart';
 
 class ViewTemplate {
-  static String viewTemplate(String name) {
+  static String viewTemplate(String name, {bool isBloc = false}) {
     final pascal = Utils.pascal(name);
+    final type = isBloc ? 'Bloc' : 'Cubit';
+    final folder = isBloc ? 'bloc' : 'cubit';
 
     return '''
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../cubit/${name}_cubit.dart';
-import '../cubit/${name}_state.dart';
+import '../$folder/${name}_$folder.dart';
+import '../$folder/${name}_state.dart';
 
 class ${pascal}View extends StatelessWidget {
   const ${pascal}View({super.key});
@@ -20,7 +22,7 @@ class ${pascal}View extends StatelessWidget {
       appBar: AppBar(
         title: const Text('$pascal'),
       ),
-      body: BlocBuilder<${pascal}Cubit, ${pascal}State>(
+      body: BlocBuilder<${pascal}$type, ${pascal}State>(
         builder: (context, state) {
           return const Center(
             child: Text('$pascal View'),